refactor: Move reusable widgets into components folder
