+++ lib/pages/main/tab_explorer/components/tick_panel.dart
@@ -1,0 +1,30 @@
+part of '../tab_explorer.dart';
+
+class _TickPanel extends StatelessWidget {
+  final String title;
+  final String contents;
+  const _TickPanel(this.title, this.contents);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(10),
+          color: LightThemeColors.cardBackground,
+        ),
+        child: Padding(
+            padding: const EdgeInsets.all(ThemePaddings.smallPadding),
+            child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  FittedBox(
+                      fit: BoxFit.scaleDown,
+                      child: Text(title, style: TextStyles.secondaryTextSmall)),
+                  FittedBox(
+                      fit: BoxFit.scaleDown,
+                      child:
+                          Text(contents, style: TextStyles.textExtraLargeBold))
+                ])));
+  }
+}
